+++ src/libs/hooks/index.js
@@ -1,0 +1,2 @@
+export { default as useTheme } from './useTheme';
+export { default as useLocale } from './useLocale';
